TSettings: add to replace buggy QSettings
Technically it's a "missing feature" rather than a "bug" since the docs aren't
incorrect.  But it's such an annoying flaw in the API that I'm calling it a
bug in the short commit summary.

See the explanation given in the doxygen comment in lib/core/TSettings.h, but
the TL;DR is that QSettings::setPath() lets us specify a **base directory**
(which will then contain a subdirectory!), instead of a filename.